+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/account/view/HistoryView.kt
@@ -7,6 +7,7 @@ import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
+import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
@@ -18,12 +19,13 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.ripple.rememberRipple
 import androidx.compose.material3.DividerDefaults
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
@@ -34,8 +36,11 @@ import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.pluralStringResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.res.vectorResource
+import androidx.compose.ui.semantics.Role
 import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
@@ -61,7 +66,11 @@ import co.electriccoin.zcash.ui.design.component.CircularMidProgressIndicator
 import co.electriccoin.zcash.ui.design.component.StyledBalance
 import co.electriccoin.zcash.ui.design.component.StyledBalanceDefaults
 import co.electriccoin.zcash.ui.design.component.TextWithIcon
+import co.electriccoin.zcash.ui.design.newcomponent.PreviewScreens
 import co.electriccoin.zcash.ui.design.theme.ZcashTheme
+import co.electriccoin.zcash.ui.design.theme.colors.ZashiColors
+import co.electriccoin.zcash.ui.design.theme.typography.ZashiTypography
+import co.electriccoin.zcash.ui.design.util.orDark
 import co.electriccoin.zcash.ui.fixture.WalletSnapshotFixture
 import co.electriccoin.zcash.ui.screen.account.HistoryTag
 import co.electriccoin.zcash.ui.screen.account.fixture.TransactionUiFixture
@@ -257,7 +266,7 @@ private fun ComposableHistoryListItemPreview() {
 }
 
 @Composable
-@Preview("History List Item Expanded")
+@PreviewScreens
 private fun ComposableHistoryListItemExpandedPreview() {
     ZcashTheme(forceDarkMode = false) {
         BlankSurface {
@@ -321,43 +330,49 @@ private fun HistoryItem(
         TransactionExtendedState.SENT -> {
             typeText = stringResource(id = R.string.account_history_item_sent)
             typeIcon = ImageVector.vectorResource(R.drawable.ic_trx_send_icon)
-            textColor = MaterialTheme.colorScheme.onBackground
-            textStyle = ZcashTheme.extendedTypography.transactionItemStyles.titleRegular
+            textColor = ZashiColors.Text.textPrimary
+            textStyle = ZashiTypography.textSm
         }
 
         TransactionExtendedState.SENDING -> {
             typeText = stringResource(id = R.string.account_history_item_sending)
             typeIcon = ImageVector.vectorResource(R.drawable.ic_trx_send_icon)
-            textColor = ZcashTheme.colors.textDescription
-            textStyle = ZcashTheme.extendedTypography.transactionItemStyles.titleRunning
+            textColor = ZashiColors.Text.textPrimary
+            textStyle = ZashiTypography.textSm
         }
 
         TransactionExtendedState.SEND_FAILED -> {
             typeText = stringResource(id = R.string.account_history_item_send_failed)
             typeIcon = ImageVector.vectorResource(R.drawable.ic_trx_send_icon)
-            textColor = ZcashTheme.colors.historyRedColor
-            textStyle = ZcashTheme.extendedTypography.transactionItemStyles.titleFailed
+            textColor = ZashiColors.Text.textError
+            textStyle =
+                ZashiTypography.textSm.copy(
+                    textDecoration = TextDecoration.LineThrough
+                )
         }
 
         TransactionExtendedState.RECEIVED -> {
             typeText = stringResource(id = R.string.account_history_item_received)
             typeIcon = ImageVector.vectorResource(R.drawable.ic_trx_receive_icon)
-            textColor = MaterialTheme.colorScheme.onBackground
-            textStyle = ZcashTheme.extendedTypography.transactionItemStyles.titleRegular
+            textColor = ZashiColors.Text.textPrimary
+            textStyle = ZashiTypography.textSm
         }
 
         TransactionExtendedState.RECEIVING -> {
             typeText = stringResource(id = R.string.account_history_item_receiving)
             typeIcon = ImageVector.vectorResource(R.drawable.ic_trx_receive_icon)
-            textColor = ZcashTheme.colors.textDescription
-            textStyle = ZcashTheme.extendedTypography.transactionItemStyles.titleRunning
+            textColor = ZashiColors.Text.textPrimary
+            textStyle = ZashiTypography.textSm
         }
 
         TransactionExtendedState.RECEIVE_FAILED -> {
             typeText = stringResource(id = R.string.account_history_item_receive_failed)
             typeIcon = ImageVector.vectorResource(R.drawable.ic_trx_receive_icon)
-            textColor = ZcashTheme.colors.historyRedColor
-            textStyle = ZcashTheme.extendedTypography.transactionItemStyles.titleFailed
+            textColor = ZashiColors.Text.textError
+            textStyle =
+                ZashiTypography.textSm.copy(
+                    textDecoration = TextDecoration.LineThrough
+                )
         }
     }
 
@@ -374,15 +389,22 @@ private fun HistoryItem(
                                     TrxItemState.EXPANDED
                                 )
                             )
+                        } else {
+                            onAction(
+                                TrxItemAction.ExpandableStateChange(
+                                    transaction.overview.rawId,
+                                    TrxItemState.COLLAPSED
+                                )
+                            )
                         }
                     }
-                    .padding(all = ZcashTheme.dimens.spacingLarge)
+                    .padding(24.dp)
                     .animateContentSize()
             )
     ) {
         Image(
             imageVector = typeIcon,
-            colorFilter = ColorFilter.tint(color = ZcashTheme.colors.secondaryColor),
+            colorFilter = ColorFilter.tint(ZashiColors.Text.textPrimary),
             contentDescription = typeText,
             modifier = Modifier.padding(top = ZcashTheme.dimens.spacingTiny)
         )
@@ -399,12 +421,10 @@ private fun HistoryItem(
                 onAction = onAction
             )
 
-            Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingXtiny))
+            Spacer(modifier = Modifier.height(2.dp))
 
             // To add an extra spacing at the end
-            Column(
-                modifier = Modifier.padding(end = ZcashTheme.dimens.spacingUpLarge)
-            ) {
+            Column {
                 val isInExpectedState =
                     transaction.expandableState == TrxItemState.EXPANDED_ADDRESS ||
                         transaction.expandableState == TrxItemState.EXPANDED_ALL
@@ -415,13 +435,13 @@ private fun HistoryItem(
                 ) {
                     HistoryItemExpandedAddressPart(onAction, transaction.recipient)
 
-                    Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingDefault))
+                    Spacer(modifier = Modifier.height(16.dp))
                 }
 
                 HistoryItemDatePart(transaction)
 
                 if (transaction.expandableState.isInAnyExtendedState()) {
-                    Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingDefault))
+                    Spacer(modifier = Modifier.height(32.dp))
 
                     HistoryItemExpandedPart(onAction, transaction)
                 }
@@ -431,7 +451,7 @@ private fun HistoryItem(
 }
 
 @Composable
-@Suppress("LongParameterList")
+@Suppress("LongParameterList", "LongMethod")
 private fun HistoryItemCollapsedMainPart(
     transaction: TransactionUi,
     typeText: String,
@@ -454,6 +474,7 @@ private fun HistoryItemCollapsedMainPart(
             text = typeText,
             style = textStyle,
             color = textColor,
+            fontWeight = FontWeight.Bold,
             modifier = Modifier.testTag(HistoryTag.TRANSACTION_ITEM_TITLE)
         )
 
@@ -468,15 +489,18 @@ private fun HistoryItemCollapsedMainPart(
         val valueTextStyle: TextStyle
         val valueTextColor: Color
         if (transaction.overview.getExtendedState().isFailed()) {
-            valueTextStyle = ZcashTheme.extendedTypography.transactionItemStyles.contentLineThrough
-            valueTextColor = ZcashTheme.colors.historyRedColor
+            valueTextStyle =
+                ZashiTypography.textSm.copy(
+                    textDecoration = TextDecoration.LineThrough
+                )
+            valueTextColor = ZashiColors.Text.textError
         } else {
-            valueTextStyle = ZcashTheme.extendedTypography.transactionItemStyles.valueFirstPart
+            valueTextStyle = ZashiTypography.textSm
             valueTextColor =
                 if (transaction.overview.isSentTransaction) {
-                    ZcashTheme.colors.historyRedColor
+                    ZashiColors.Text.textError
                 } else {
-                    ZcashTheme.colors.textPrimary
+                    ZashiColors.Text.textPrimary
                 }
         }
 
@@ -500,7 +524,7 @@ private fun HistoryItemCollapsedMainPart(
             textStyle =
                 StyledBalanceDefaults.textStyles(
                     mostSignificantPart = valueTextStyle,
-                    leastSignificantPart = ZcashTheme.extendedTypography.transactionItemStyles.valueSecondPart
+                    leastSignificantPart = ZashiTypography.textXxs
                 ),
             textColor = valueTextColor,
         )
@@ -545,8 +569,8 @@ private fun HistoryItemCollapsedAddressPart(
 
                 Text(
                     text = transaction.recipient.addressValue,
-                    style = ZcashTheme.extendedTypography.transactionItemStyles.addressCollapsed,
-                    color = ZcashTheme.colors.textDescription,
+                    style = ZashiTypography.textSm,
+                    color = ZashiColors.Text.textTertiary,
                     maxLines = 1,
                     overflow = TextOverflow.Ellipsis,
                     modifier =
@@ -580,26 +604,29 @@ private fun HistoryItemExpandedAddressPart(
     ) {
         Text(
             text = recipient.addressValue,
-            style = ZcashTheme.extendedTypography.transactionItemStyles.content,
-            color = ZcashTheme.colors.textPrimary,
+            style = ZashiTypography.textSm,
+            color = ZashiColors.Text.textTertiary,
             modifier =
                 Modifier
                     .fillMaxWidth(EXPANDED_ADDRESS_WIDTH_RATIO)
         )
 
-        Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingTiny))
+        Spacer(modifier = Modifier.height(16.dp))
 
         TextWithIcon(
             text = stringResource(id = R.string.account_history_item_tap_to_copy),
-            style = ZcashTheme.extendedTypography.transactionItemStyles.content,
-            color = ZcashTheme.colors.textDescription,
+            style = ZashiTypography.textSm,
+            color = ZashiColors.Btns.Tertiary.btnTertiaryFg,
+            fontWeight = FontWeight.SemiBold,
             iconVector = ImageVector.vectorResource(R.drawable.ic_trx_copy),
-            iconTintColor = ZcashTheme.colors.secondaryColor,
+            iconTintColor = ZashiColors.Text.textTertiary,
             modifier =
                 Modifier
-                    .clip(RoundedCornerShape(ZcashTheme.dimens.regularRippleEffectCorner))
-                    .clickable { onAction(TrxItemAction.AddressClick(recipient)) }
-                    .padding(all = ZcashTheme.dimens.spacingTiny)
+                    .clickable(
+                        role = Role.Button,
+                        indication = rememberRipple(radius = 2.dp, color = ZashiColors.Text.textTertiary),
+                        interactionSource = remember { MutableInteractionSource() }
+                    ) { onAction(TrxItemAction.AddressClick(recipient)) }
         )
     }
 }
@@ -619,8 +646,8 @@ private fun HistoryItemDatePart(
     if (formattedDate != null) {
         Text(
             text = formattedDate,
-            style = ZcashTheme.extendedTypography.transactionItemStyles.content,
-            color = ZcashTheme.colors.textDescription,
+            style = ZashiTypography.textSm,
+            color = ZashiColors.Text.textTertiary,
             maxLines = 1,
             overflow = TextOverflow.Ellipsis,
             modifier = modifier
@@ -643,8 +670,8 @@ private fun HistoryItemExpandedPart(
                         id = R.plurals.account_history_item_message,
                         count = transaction.messages!!.size
                     ),
-                style = ZcashTheme.extendedTypography.transactionItemStyles.contentMedium,
-                color = ZcashTheme.colors.textPrimary
+                style = ZashiTypography.textSm,
+                color = ZashiColors.Text.textTertiary,
             )
 
             Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingSmall))
@@ -667,45 +694,19 @@ private fun HistoryItemExpandedPart(
         ) {
             Text(
                 text = stringResource(id = R.string.account_history_item_no_message),
-                style = ZcashTheme.extendedTypography.transactionItemStyles.contentItalic,
-                color = ZcashTheme.colors.textPrimary,
+                style = ZashiTypography.textSm,
+                color = ZashiColors.Text.textTertiary,
                 modifier = Modifier.fillMaxWidth(EXPANDED_TRANSACTION_WIDTH_RATIO)
             )
 
             Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingDefault))
         }
 
-        HistoryItemTransactionIdPart(
-            transaction = transaction,
-            onAction = onAction
-        )
+        HistoryItemTransactionIdPart(transaction = transaction, onAction = onAction)
 
-        Spacer(modifier = (Modifier.height(ZcashTheme.dimens.spacingDefault)))
+        Spacer(modifier = Modifier.height(16.dp))
 
         HistoryItemTransactionFeePart(fee = transaction.overview.feePaid)
-
-        Spacer(modifier = (Modifier.height(ZcashTheme.dimens.spacingLarge)))
-
-        TextWithIcon(
-            text = stringResource(id = R.string.account_history_item_collapse_transaction),
-            style = ZcashTheme.extendedTypography.transactionItemStyles.contentUnderline,
-            color = ZcashTheme.colors.textDescription,
-            iconVector = ImageVector.vectorResource(id = R.drawable.ic_trx_collapse),
-            modifier =
-                Modifier
-                    .clip(RoundedCornerShape(ZcashTheme.dimens.regularRippleEffectCorner))
-                    .clickable {
-                        if (transaction.expandableState >= TrxItemState.EXPANDED) {
-                            onAction(
-                                TrxItemAction.ExpandableStateChange(
-                                    transaction.overview.rawId,
-                                    TrxItemState.COLLAPSED
-                                )
-                            )
-                        }
-                    }
-                    .padding(all = ZcashTheme.dimens.spacingTiny)
-        )
     }
 }
 
@@ -735,35 +736,38 @@ private fun HistoryItemTransactionIdPart(
         ) {
             Text(
                 text = stringResource(id = R.string.account_history_item_transaction_id),
-                style = ZcashTheme.extendedTypography.transactionItemStyles.content,
-                color = ZcashTheme.colors.textDescription,
+                style = ZashiTypography.textSm,
+                color = ZashiColors.Text.textTertiary,
             )
 
-            Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingTiny))
+            Spacer(modifier = Modifier.height(8.dp))
 
             Text(
                 text = txIdString,
-                style = ZcashTheme.extendedTypography.transactionItemStyles.content,
-                color = ZcashTheme.colors.textPrimary,
+                style = ZashiTypography.textSm,
+                color = ZashiColors.Text.textTertiary,
                 modifier =
                     Modifier
                         .fillMaxWidth(EXPANDED_TRANSACTION_WIDTH_RATIO)
                         .testTag(HistoryTag.TRANSACTION_ID)
             )
 
-            Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingTiny))
+            Spacer(modifier = Modifier.height(16.dp))
 
             TextWithIcon(
                 text = stringResource(id = R.string.account_history_item_tap_to_copy),
-                style = ZcashTheme.extendedTypography.transactionItemStyles.content,
-                color = ZcashTheme.colors.textDescription,
+                style = ZashiTypography.textSm,
+                color = ZashiColors.Btns.Tertiary.btnTertiaryFg,
+                fontWeight = FontWeight.SemiBold,
                 iconVector = ImageVector.vectorResource(R.drawable.ic_trx_copy),
-                iconTintColor = ZcashTheme.colors.secondaryColor,
+                iconTintColor = ZashiColors.Text.textTertiary,
                 modifier =
                     Modifier
-                        .clip(RoundedCornerShape(ZcashTheme.dimens.regularRippleEffectCorner))
-                        .clickable { onAction(TrxItemAction.TransactionIdClick(txIdString)) }
-                        .padding(all = ZcashTheme.dimens.spacingTiny)
+                        .clickable(
+                            role = Role.Button,
+                            indication = rememberRipple(radius = 2.dp, color = ZashiColors.Text.textTertiary),
+                            interactionSource = remember { MutableInteractionSource() }
+                        ) { onAction(TrxItemAction.TransactionIdClick(txIdString)) }
             )
         } else {
             Row(
@@ -782,21 +786,21 @@ private fun HistoryItemTransactionIdPart(
                                 )
                             )
                         }
-                        .padding(all = ZcashTheme.dimens.spacingTiny)
             ) {
                 Text(
                     text = stringResource(id = R.string.account_history_item_transaction_id),
-                    style = ZcashTheme.extendedTypography.transactionItemStyles.content,
-                    color = ZcashTheme.colors.textDescription,
+                    style = ZashiTypography.textSm,
+                    color = ZashiColors.Text.textTertiary,
                 )
 
-                Spacer(modifier = Modifier.width(ZcashTheme.dimens.spacingSmall))
+                Spacer(modifier = Modifier.weight(1f))
 
                 Text(
                     text = txIdString,
-                    style = ZcashTheme.extendedTypography.transactionItemStyles.content,
-                    color = ZcashTheme.colors.textDescription,
+                    style = ZashiTypography.textSm,
+                    color = ZashiColors.Text.textTertiary,
                     maxLines = 1,
+                    textAlign = TextAlign.End,
                     overflow = TextOverflow.Ellipsis,
                     modifier =
                         Modifier
@@ -813,14 +817,14 @@ private fun HistoryItemTransactionFeePart(
     fee: Zatoshi?,
     modifier: Modifier = Modifier
 ) {
-    Column(modifier = modifier) {
+    Row(modifier = modifier) {
         Text(
             text = stringResource(id = R.string.account_history_item_transaction_fee),
-            style = ZcashTheme.extendedTypography.transactionItemStyles.content,
-            color = ZcashTheme.colors.textDescription,
+            style = ZashiTypography.textSm,
+            color = ZashiColors.Text.textTertiary,
         )
 
-        Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingTiny))
+        Spacer(modifier = Modifier.weight(1f))
 
         if (fee == null) {
             Text(
@@ -829,8 +833,8 @@ private fun HistoryItemTransactionFeePart(
                         id = R.string.account_history_item_transaction_fee_typical,
                         DEFAULT_FEE
                     ),
-                style = ZcashTheme.extendedTypography.transactionItemStyles.feeFirstPart,
-                color = ZcashTheme.colors.textDescription,
+                style = ZashiTypography.textSm,
+                color = ZashiColors.Text.textTertiary,
             )
         } else {
             StyledBalance(
@@ -839,10 +843,10 @@ private fun HistoryItemTransactionFeePart(
                 isHideBalances = false,
                 textStyle =
                     StyledBalanceDefaults.textStyles(
-                        mostSignificantPart = ZcashTheme.extendedTypography.transactionItemStyles.feeFirstPart,
-                        leastSignificantPart = ZcashTheme.extendedTypography.transactionItemStyles.feeSecondPart
+                        mostSignificantPart = ZashiTypography.textSm,
+                        leastSignificantPart = ZashiTypography.textXxs
                     ),
-                textColor = ZcashTheme.colors.textDescription
+                textColor = ZashiColors.Text.textTertiary
             )
         }
     }
@@ -858,11 +862,14 @@ private fun HistoryItemMessagePart(
     val textStyle: TextStyle
     val textColor: Color
     if (state.isFailed()) {
-        textStyle = ZcashTheme.extendedTypography.transactionItemStyles.contentLineThrough
-        textColor = ZcashTheme.colors.historyRedColor
+        textStyle =
+            ZashiTypography.textSm.copy(
+                textDecoration = TextDecoration.LineThrough
+            )
+        textColor = ZashiColors.Text.textError
     } else {
-        textStyle = ZcashTheme.extendedTypography.transactionItemStyles.content
-        textColor = ZcashTheme.colors.textPrimary
+        textStyle = ZashiTypography.textSm
+        textColor = ZashiColors.Text.textPrimary
     }
 
     Column(modifier = modifier.then(Modifier.fillMaxWidth())) {
@@ -870,12 +877,12 @@ private fun HistoryItemMessagePart(
         val bubbleStroke: BorderStroke
         val arrowAlignment: BubbleArrowAlignment
         if (state.isSendType()) {
-            bubbleBackgroundColor = Color.Transparent
-            bubbleStroke = BorderStroke(1.dp, ZcashTheme.colors.textFieldFrame)
+            bubbleBackgroundColor = ZashiColors.Utility.Gray.utilityGray200 orDark Color.Transparent
+            bubbleStroke = BorderStroke(1.dp, ZashiColors.Text.textPrimary)
             arrowAlignment = BubbleArrowAlignment.BottomLeft
         } else {
-            bubbleBackgroundColor = ZcashTheme.colors.historyMessageBubbleColor
-            bubbleStroke = BorderStroke(1.dp, ZcashTheme.colors.historyMessageBubbleStrokeColor)
+            bubbleBackgroundColor = ZashiColors.Utility.Gray.utilityGray200 orDark Color.Transparent
+            bubbleStroke = BorderStroke(1.dp, ZashiColors.Text.textPrimary)
             arrowAlignment = BubbleArrowAlignment.BottomRight
         }
 
@@ -893,19 +900,23 @@ private fun HistoryItemMessagePart(
             )
         }
 
-        Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingTiny))
+        Spacer(modifier = Modifier.height(16.dp))
 
         TextWithIcon(
             text = stringResource(id = R.string.account_history_item_tap_to_copy),
-            style = ZcashTheme.extendedTypography.transactionItemStyles.content,
-            color = ZcashTheme.colors.textDescription,
+            style = ZashiTypography.textSm,
+            color = ZashiColors.Btns.Tertiary.btnTertiaryFg,
+            fontWeight = FontWeight.SemiBold,
             iconVector = ImageVector.vectorResource(R.drawable.ic_trx_copy),
-            iconTintColor = ZcashTheme.colors.secondaryColor,
+            iconTintColor = ZashiColors.Text.textTertiary,
             modifier =
                 Modifier
-                    .clip(RoundedCornerShape(ZcashTheme.dimens.regularRippleEffectCorner))
-                    .clickable { onAction(TrxItemAction.MessageClick(message)) }
-                    .padding(all = ZcashTheme.dimens.spacingTiny)
+                    .clickable(
+                        onClick = { onAction(TrxItemAction.MessageClick(message)) },
+                        role = Role.Button,
+                        indication = rememberRipple(radius = 2.dp, color = ZashiColors.Text.textTertiary),
+                        interactionSource = remember { MutableInteractionSource() }
+                    )
         )
     }
 }